Improved initialisation of word vector models, added .vec embeddings file
