+++ server/suggest/spell_checker.py
@@ -1,0 +1,4 @@
+"""
+Uses word embedding models to check the spelling of words and suggest corrections.
+"""
+
