+++ server/users/session.py
@@ -6,7 +6,7 @@ from sanic.request import Request
 from server.app import BaseModel
 from server.mongo.document import Document
 
-from server.word_embedding.supervised_models import load_model, SupervisedModels
+from server.word_embedding.sanic_supervised_models import load_model, SupervisedModels
 
 
 class Session(BaseModel, Document):
